+++ back/src/main/java/com/rdrg/back/service/FileService.java
@@ -1,0 +1,11 @@
+package com.rdrg.back.service;
+
+import org.springframework.core.io.Resource;
+import org.springframework.web.multipart.MultipartFile;
+
+public interface FileService {
+    
+    String  upload(MultipartFile file);
+    Resource GetFile(String fileName);
+
+}
