+++ back/src/main/java/com/rdrg/back/service/implementation/BoardServiceImplementation.java
@@ -1,13 +1,9 @@
 package com.rdrg.back.service.implementation;
 
 import java.util.List;
-import java.io.File;
-import java.util.UUID;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.rdrg.back.dto.request.board.PostBoardRequestDto;
 import com.rdrg.back.dto.request.board.PostCommentRequestDto;
@@ -47,44 +43,6 @@ public class BoardServiceImplementation implements BoardService {
         }
 
         return ResponseDto.success();
-
-    }
-
-    @Value("${file.url}")private String fileUrl;
-    @Value("${file.path}")private String filePath;
-
-    @Override
-    public String upload(MultipartFile file) {
-
-        // 빈파일인지 검증
-        if(file.isEmpty()) return null;
-
-        // 원본 파일명  가져오기
-        String originalFileName = file.getOriginalFilename();
-        // 원본 파일명 확장자를 구함
-        String extension = originalFileName.substring(originalFileName.lastIndexOf("."));
-        // UUID 혁식의 임의의 문자열 생성
-        String uuid = UUID.randomUUID().toString();
-        // 저장시 사용할 파일명 생셩()
-        String saveFileName = uuid + extension;
-        // 저장할 경로 생성
-        String savePath = filePath + saveFileName;
-
-        try {
-
-            // 파일 저장
-            file.transferTo(new File(savePath));
-
-        } catch (Exception exception) {
-
-            exception.printStackTrace();
-            return null;
-
-        }
-
-        // 파일을 불러올 수 있는 경로 생성
-        String url = fileUrl + saveFileName;
-        return url;
 
     }
